+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## vNext
 
 * Adding compression middleware. [https://github.com/loco-rs/loco/pull/205](https://github.com/loco-rs/loco/pull/205)
+  Added support for [compression middleware](https://docs.rs/tower-http/0.5.0/tower_http/compression/index.html).
+  usage:
+
+```yaml
+middlewares:
+  compression:
+    enable: true
+```
 * Create a new Database from the CLI. [https://github.com/loco-rs/loco/pull/223](https://github.com/loco-rs/loco/pull/223)
 * Validate if seaorm CLI is installed before running `cargo loco db entities` and show a better error to the user. [https://github.com/loco-rs/loco/pull/212](https://github.com/loco-rs/loco/pull/212)
 * Adding to `saas and `rest-api` starters a redis and DB in GitHub action workflow to allow users work with github action out of the box. [https://github.com/loco-rs/loco/pull/215](https://github.com/loco-rs/loco/pull/215)