+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNext
 
+* Adding [pagination](https://loco.rs/docs/the-app/pagination/) on Models. [https://github.com/loco-rs/loco/pull/238](https://github.com/loco-rs/loco/pull/238)
 * Adding compression middleware. [https://github.com/loco-rs/loco/pull/205](https://github.com/loco-rs/loco/pull/205)
   Added support for [compression middleware](https://docs.rs/tower-http/0.5.0/tower_http/compression/index.html).
   usage: